+++ src/project1/src/project1/search_node.cpp
@@ -6,6 +6,6 @@
 #include <algorithm>
 
 namespace project1 {
-  SearchNode::SearchNode( const double& xArg, const double& yArg , const double& thetaArg, const double& fArg, const double& gArg, const double& hArg , const std::shared_ptr< SearchNode >& bpArg ) : x(xArg), y(yArg), theta(thetaArg), f(fArg), g(gArg), h(hArg), bp(bpArg) {}
+  SearchNode::SearchNode( const int& xArg, const int& yArg , const double& thetaArg, const double& fArg, const double& gArg, const double& hArg , const std::shared_ptr< SearchNode >& bpArg ) : x(xArg), y(yArg), theta(thetaArg), f(fArg), g(gArg), h(hArg), bp(bpArg) {}
 
 }
